==> data/scripts/dense/3a_sentence_embeddings.ipynb <==
This notebook uses BERT to extract sentence embeddings from the memory descriptions.

Following this tutorial: https://www.geeksforgeeks.org/how-to-generate-word-embedding-using-bert/

import numpy as np
import pandas as pd
import random
import torch
from transformers import BertTokenizer, BertModel
from sklearn.metrics.pairwise import cosine_similarity
import matplotlib.pyplot as plt
import scipy.stats
from sklearn.cluster import KMeans
from sklearn.decomposition import PCA
from sklearn.metrics import silhouette_score

Set random seed

random_seed = 9
random.seed(random_seed)

# set for PyTorch
torch.manual_seed(random_seed)
if torch.cuda.is_available():
    torch.cuda.manual_seed_all(random_seed)

Load BERT

tokenizer = BertTokenizer.from_pretrained('bert-base-uncased')
model = BertModel.from_pretrained('bert-base-uncased')
# gave a JS error before, but it went away... (if happens again, ctrl+refresh button)
# https://stackoverflow.com/questions/73715821/jupyter-lab-issue-displaying-widgets-javascript-error

# Sentence embeddings for all memories

First, we generate sentence embeddings for all memories reported. While this is not used specifically in the content analysis (comparing across pairs of memories), it could be useful in future exploratory analysis.

**Note:** we don't filter out "private" memories!

### Load the data

all_memories = pd.read_csv('../../processed/dense/memory/all_memories.csv', header = 0)

all_memories

Only keep the identifying columns and the memory description.

all_memories = all_memories[['internal_id', 'song_id', 'orig_or_cover', 'description']]

all_memories

### Generate sentence embeddings

sentence_embeddings = []

for row in range(#len(all_memories)):
    # print every fifty rows to keep track of where we are
    if row % 50 == 0: print(row)
    
    # grab this row - keep names consistent with dataframe
    internal_id = all_memories.iloc[row]["internal_id"]
    song_id = all_memories.iloc[row]["song_id"]
    orig_or_cover = all_memories.iloc[row]["orig_or_cover"]
    description = all_memories.iloc[row]["description"]

    # tokenize the input text for the memory description
    encoding = tokenizer.batch_encode_plus([description], padding=True, truncation=True, return_tensors='pt', add_special_tokens=True)
    input_ids = encoding['input_ids']
    attention_mask = encoding['attention_mask']
    
    # generate embeddings
    with torch.no_grad():
        outputs = model(input_ids, attention_mask=attention_mask)
        sentence_embedding = outputs.pooler_output
        #average_word_embedding = outputs.last_hidden_state.mean(dim=1) # if we want to compare average word embeddings to pooled output

    # append the embeddings to our list
    sentence_embeddings.append([internal_id, song_id, orig_or_cover] + sentence_embedding.tolist()[0])

# turn list of embeddings into a dataframe
embeddings_df = pd.DataFrame(sentence_embeddings, columns = ['internal_id', 'song_id', 'orig_or_cover'] + np.arange(1,769).tolist())
embeddings_df

(Fifty rows takes about 2 seconds)

Save the output.

embeddings_df.to_csv('../../processed/dense/memory/all_memory_embeddings.csv', index = False)

# Semantic distance for memory pairs

Now that we have all of the sentence embeddings, we compute semantic distance between pairs of memories using cosine similarity.

First, identify the pairs.

embedding_distances = []

for sub in pd.unique(embeddings_df['internal_id']):
#for sub in [1,2]:
    # get just this participant's memories
    this_sub = embeddings_df[embeddings_df['internal_id'] == sub]

    # get the songs and their counts 
    these_songs, these_song_counts = np.unique(this_sub['song_id'], return_counts=True)

    # if the song shows up twice, then that's a memory pair
    for i in range(len(these_songs)):
        if these_song_counts[i] == 2: 
            song_id = these_songs[i]
            this_song = this_sub[this_sub['song_id'] == song_id]
            orig_embedding = this_song[this_song['orig_or_cover'] == 'orig'].to_numpy()[0,3:]
            cover_embedding = this_song[this_song['orig_or_cover'] == 'cover'].to_numpy()[0,3:]
            # compute semantic distance using Euclidean distance
            semantic_distance = np.linalg.norm(orig_embedding - cover_embedding)

            embedding_distances.append([sub, song_id, semantic_distance])
            
embedding_dist_df = pd.DataFrame(embedding_distances, columns = ['internal_id', 'song_id', 'semantic_distance'])
embedding_dist_df

print(np.max(embedding_dist_df['semantic_distance']))

embedding_dist_df.to_csv('../../processed/dense/memory/memory_pairs_semantic_distances.csv', index = False)
